movie_review_api/movie_review_api/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
gunicorn==22.0.0
whitenoise==6.7.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
